requirements.txt
Flask>=0.12.2
Flask-SQLAlchemy>=2.2
PyMySQL>=0.7.11
Pillow>=4.2.1
